backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
beanie==1.23.6
motor==3.3.1
pydantic==2.4.2
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
typing-extensions==4.8.0
starlette==0.27.0
groq==0.4.2 